fix: apply requested changes
